pyproject.toml
[project]
name = "notification-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "aiosmtplib>=5.0.0",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "celery>=5.6.2",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "flower>=2.0.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
asyncio_default_test_loop_scope = "function"
